requirements.txt
altair==4.2.0
attrs==22.1.0
backports.zoneinfo==0.2.1
blinker==1.5
blis==0.7.9
breadability==0.1.20
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==2.1.1
commonmark==0.9.1
confection==0.0.3
contourpy==1.0.6
cycler==0.11.0
cymem==2.0.7
decorator==5.1.1
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
entrypoints==0.4
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.29
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
langcodes==3.3.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.2
murmurhash==1.0.9
nltk==3.7
numpy==1.23.5
packaging==21.3
pandas==1.5.2
pathy==0.10.0
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
plac==1.1.3
plotly==5.11.0
preshed==3.0.8
protobuf==3.20.3
pyarrow==10.0.1
pycountry==22.3.5
pydantic==1.8.2
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.1
rich==12.6.0
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
semver==2.13.0
six==1.16.0
sklearn==0.0.post1
smart-open==5.2.1
smmap==5.0.0
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.5
streamlit==1.15.2
sumy==0.11.0
tenacity==8.1.0
thinc==8.1.5
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typer==0.7.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
wasabi==0.10.1
watchdog==2.2.0
wincertstore==0.2
zipp==3.11.0
textblob==0.17.1
